Update idempotence test
The idempotence test will now try to leverage the idempotence ansible
callback plugin to determine whether the role is idempotent and if not, display
the tasks that are not.

In case the callback plugin is not found, the current behavior is not
altered and only a success or failure message is displayed to the user.

The ansible idempotence callback plugin was included into the molecule
package.

Add unit tests for the idempotence tests.

Update setup.cfg.