Ansible inventory SSH options for better testinfra support
Testinfra doesn't inherit all of the SSH options that Molecule uses during the converge phase which was causing testinfra to not be able to run because it couldn't SSH to the VM. It was prompting to accept the SSH fingerprint and would just hang. Adding SSH's StrictHostKeyChecking=no and UserKnownHostsFile=/dev/null to the ansible_inventory file allows testinfra to properly SSH to the VM and perform the tests.