Add support for group_vars
It happens that for testing some roles in a specific context, you want
to use group_vars instead of maintaining copies of the variables in the
molecule playbook.

This patch adds a configuration option in the molecule section allowing
you to specify the path of the group_vars to use.