Add tab completion support
This patch provides a script to add tab completion support to molecule.
It auto-completes the subcommands, options and dynamic arguments such as
platform, providers, and hosts.

To ease tabular display and porcelain mode, a new
``_display_tabulate_data`` function was implemented in the Molecule
object.

The documentation was updated accordingly.