Corrected `molecule init`
Corrected bug introduced in #268.  Removed duplicate "are we running
the init command" logic from both the core and command base class.

The `init` sub command is unique to other commands.  It doesn't require
a `molecule.yml`.  Moved this logic into a method, in the parent class,
so the children can change the behavior.

Also disabled static inventory introduced in #40.  We never used this,
and the core needs refactored before we attempt to reintroduce it.

Fixes: #275