Added a spec_helper which supports docker (#327)
Since serverspec is not the default test framework, we do not supply
helper files.  We simply use rake to act as an executer.  This will
serve as an example implementation for others.

Special thanks to @kireledan, who actually figured this out, but I sent
him on a wild goose hunt, thinking we needed to change the Rakefile.

Fixes: #272