Renamed provisioner to driver
So we can start making changes to the core config syntax (#204), we should
start calling the provisioner as driver.  Renamed all mentions of
provisioner to driver.

This is also the beginning of renaming our directories to non-plural.
Will submit a follow-up pull request once the outstanding PRs land.

This is a non-breaking config change.