Add support for molecule generated keypair and keyfile, if omitted in the molecule.yml
This change is so that molecule can generate an ssh keyfile and openstack keypair when not specified in the molecule.yml
The reason for this change, is we plan to use this mostly in continuous integration systems. We don't really care about persisting a key or keypair since the instance running molecule will be destroyed once molecule is finished.
The changes make sure that if ~/.ssh/id_rsa exists already, that we will not delete it. Molecule only deletes the keypair and keyfiles that it creates and should not do so if they are explictly specified.