Removed customconf from state file
The state file was used to determine if ansible.cfg was custom or
provided by the user.  However, the logic was flawed.  We execute
`destroy` which first executes `remove_templates`.  Unfortunately,
`remove_templates` relies on the state file having `customconf` set,
but the setting occurs in another method, which is called after
`destroy`.

We now look for a "Molecule managed" string, to know if safe to delete
or not.