Enhancement #463 - Convert vagrantfile from relying on jinja
Reworked the generation of the Vagrantfile to use Ruby in the file itself
instead of Jinja2. The new flow matches that of how the Serverspec file
is generated in that needed configuration parameters are written to the
state file and cookiecutter is used to pass the state file into the
Vagrant file for inclusion. From there, logic in the Vagrantfile itself
manages the configuration.