Correct too many authentication failures error
By default, SSH will try to offer all keys found under ~/.ssh. If it
reaches the limit of authentication failures (default=6), the connection
will be closed. This is a problem for people with many SSH keys. To work
around this issue and only use the specified IdentifyFile (-i), the
option IdentitiesOnly should be specified.

Fixes: #749