Initial work of static driver
This driver allows one to supply externally managed instances to
molecule.  Molecule skips the create/destroy playbooks on this driver.
It is up to the user to properly configure ansible connect to these
systems.

It should be noted, one cannot connect externally to lxc, lxd, or docker
when using their respectively named `ansible_connection`.  The
containers would need to be running ssh for this to be possible.

`molecule test` runs successfully against local statically managed
docker instances.  Static Docker functional tests are passing as well.