Reworked scenario name
The scenario name cannot be configured through`molecule.yml`.
It is determined by the directory name of the scenario.  The
config value references a `MOLECULE_SCENARIO_NAME` environment
variable, which is correctly interpolated.  However, it provided
so that the config is symmetrical.  It may be removed in the future,
if the source of much confusion.