Add yamllint as a linter
Added yamllint as a linter that always run prior to Ansible Lint.
This acts much like our flake8 verifier with Testinfra.  There is
no way to disable the yamllint linting, unless you also disable
the Ansible Lint linter.

I'm not 100% happy with this implementation, since it feels a bit
gross overloading Ansible Lint with yamllint.  However, we do
this with Testinfra already.

Fixes: #869