Append path where dependencies are installed
If a user uses dependencies they are installed into .molecule/roles/.
The path was not updated to reflect this.  Also updated library and
plugin path to a similar pattern.  Granted, most users will not do this,
since galaxy sucks for sharing libraries and plugins.  However, gilt
works with this workflow well. *plug*