Absolute path ANSIBLE vars
Passing relative paths gets complicated based on the consumer
and type.

As an example:

The `ansible-lint` command uses `ANSIBLE_ROLES_PATH` relative to
where the command is run.  However, with `ansible-playbook` it is
relative to the inventory file.

Making matters worse.

Ansible doesn't treat ANSIBLE_LIBRARY relative pathing the same as
it does with `ANSIBLE_ROLES_PATH`.  Anyways, this is annoying.
Using full paths solves this issue.