Add cleanup step.
Adds a cleanup step to Molecule which is intended to be used to
undo changes made during a prepare playbook. If one has to provision
resources outside of the Molecule environment one can use this
cleanup step to ensure those resources are cleaned up after Molecule
is done executing.

The cleanup step runs directly before the destroy step.

Related to #1638

Signed-off-by: Nick Wilburn <senior.crepe@gmail.com>