allow overriding pytest functional arguments with posargs
It is useful to run a single test from the command line with, for instance:

  tox -e ansibledevel-functional -- -k 'test_command_init_role[docker]' test/functional/test_command.py

The test/functional/ pytest argument is made a default used when there
are no posargs instead of being hardcoded and always used. If all
pytest invocations include test/functional/ there is no way for the
user to restrict the tests being run to a single file.

Signed-off-by: singuliere <singuliere@autistici.org>