Make default schenario name use the parent folder
Changes the default scenario name from being just hardcoded 'default'
to be the directory name hosting the molecule.yml file.

This change will allow us to reuse (symlink) molecule.yml file across
scenarios and roles because we will not longer be forced to define
this field in each of them.

This also fixes bug where it was not possible to do:
	molecule init scenario -s foo

Now if the role [-r/--role] is not mentioned, it is infered from the
current directory. Documentation states that init a scenario is to
run from inside the role directory.

This change is safe because it makes the role parameter optional instead
of mandatory, not changing its old behavior.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>
Related-Bug: #1744