Adapt functional testing to recent lint updates
Starting in ansible-lint 4.0, the role metadata is now being linted
to ensure the metadata file has been updated by the role author prior
to submission to ansible galaxy. This is a "good thing", but
unfortunately breaks molecule CI.

This solution attempts to retain the behavior of reporting these lint
errors to users, but add behavior to functional tests to independently
lint and fix this file prior to testing it with molecule.

I have also updated references to the cookiecutter template to use the
new location in the github 'ansible' namespace.

Signed-off-by: Sean Myers <sean.myers@redhat.com>