DRY up sequence command execution
Removes a lot of repeated boilerplate found in commands that run
scenario sequences, centralizing it in the base command module.

Aside from removing duplicated code, this should also help make it a
little easier to do introspection during sequence execution and take
actions for all sequences based on the config, as seen with the
'destroy==always' handling. That behavior could now be added to any
sequence by adding support for the `--destroy` option to the decorated
command function.

Signed-off-by: Sean Myers <sean.myers@redhat.com>