Scenarios can now prune themselves
This decouples scenario pruning from the `destroy` command, allowing
scenarios to be pruned in any context where a scenario instance is
available.

This slightly modifies the pruning behavior. Previously this would
unconditionally prune whenever a 'destroy' sequence step was executed.
Now, `prune` is only called at the end of a sequence if that sequence
included a `destroy` sequence step. This prevents unexpected pruning
during sequence execution. As an example, users attempting to use
ansible galaxy roles in the `create` and `destroy` steps would need to
very carefully ensure the `dependency` step is run before *and after*
every destroy step. With this change, a single dependency call is
needed.

Signed-off-by: Sean Myers <sean.myers@redhat.com>