Avoid using incompatible python with doc
Not mentioning basepython on tox doc environment risks allowing it
to use py27 which will translate to instant failure while encountering
the f'' strings.

Also using a 'python3' value for basepython would produce failures
on some systems where tox will pick python3.4 and use it, failing
while installing dependencies.

The only known way to avoid this issue with tox is to mention a
hardcoded full version of python which is known as supported, assuring
that we control the experience.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>