Unpin molecule requirements
Also adds a `pip check` execution for testing that there are no
conflicts after instalation of all requirements.

Temporary the exit code returned by `pip check` is ignored as we have
a known conflict caused by `ansible[azure]` dependency.

Avoids installing lxc on MacOS(Darwin) to avoid test failures as lxc
does not even compile on that platform.

Forces pytest<4.0.0 because one plugin does not support it yet but
includes comment so we can fix that as soon the plugin is updated.

Fixes: #1831
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>