Use Ansible checks for functional driver support
Since this tool is tightly coupled to Ansible, it makes sense to make
use of their dependency checks for the purpose of skipping tests.
Otherwise, we make a check against only the environment variables that
we need to make our tests run. In the case Ansible does not provide us
with a specific import, we do it ourself (e.g, shade).

Allows developers install Docker and run run `tox -e functional` without
issues of lacking credentials for the various drivers.

Signed-off-by: Luke Murphy <lukewm@riseup.net>