fix: EC2 create security group in correct vpc (#1644)
Currently the security group is created in the default vpc. If we provide
a non-default vpc_subnet_id for the Molecule instance the security group
and instance will be in different vpcs and the instance creation will
fail with:
```
"msg": "The following group names are not valid: molecule"
```

Patch the module to create the security group in the same subnet as the
instance. This is done for each instance mentioned in molecule.yml
platforms.

Signed-off-by: Dani Hodovic <dani.hodovic@gmail.com>