Execute cleanup on lint failure (#1902)
Errors during linting automatically called the destroy playbook,
bypassing the cleanup if it is defined. This commit adds a test that
validates that the cleanup step is properly called and calls the task
directly.

This could be improved by somehow linking the destroy task to the
cleanup task, making sure that calling destroy always calls cleanup
first it is defined.

Signed-off-by: Ghislain Bourgeois <ghislain.bourgeois@gmail.com>