Pin yapf to <0.27 (0.26)
Format checking is failing with yapf 0.27, but if yapf 0.27's changes are
applied, linting fails. We need to deal with that, but CI should not
continue to be broken while that discussion takes place, so I suggest
pinning yapf to keep CI working while we come to a consensus on what to
do about yapf.

Signed-off-by: Sean Myers <sean.myers@redhat.com>