Increase yamllint coverage
Makes yamllinting execution to cover files that previously were not
linted in order to assure consistent formatting.

This also enables people to run yamllint without extra parameters
and avoid getting errors.

Contains only one exception caused by a bug in pyup tool but we
referenced the bug and once is fixed we can remove it.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>