configure scenario action in execute_subcommand (#1913)
The ansible-playbook provisioner uses the config.action to decide
whether or not to pass provisioner args to sequence steps. By default it
does not do this for `create` and `destroy` steps. In the case where
`destroy` (and now also `cleanup`) is called as a result of an error in
a test sequence, molecule was not setting the scenario's config.action
correctly. My recent changes to the command runners didn't address this
problem, but it should be a irelatively simple fix now that they're applied:

Centralize the setting of the config action into
`molecule.command.base.execute_subcommand`.

Signed-off-by: Sean Myers <sean.myers@redhat.com>