Enable use of pre-commit for linting
pre-commit is framework used to orchestrating execution of many small
utilities (usually linters) which is also used by tox itself and should
not be seen as a replacement for it.

Despite its name pre-commit does not require installation of a git hook
in order to be used, this being an optional feature of the tool.

Among benefits we can count:

* lower disk footprint as cached linters are reused across repos
* must faster (even on CI due to ~/.cache/pre-commit)
* avoids conflicts between linters, each of them being isolated
* can run only changed files (w/o --all-files flag)
* pinned dependencies by design
* one line bumping support `pre-commit autoupdate`
* optional git hook (pre-commit install)
* allows user to run a single linter if they want

This change does not require any workflow changes for developers, as
`tox -e lint` will still run the linters.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>