Use selinux shim library for testing
Reverst the use of site-packages when testing with tox as this proved
to cause other issues and conflicts and use selinux shim module in order
to allow ansible to work correctly from inside the virtualenv.

It should be noted that this is more of a indirect and weak-runtime
dependency due to use of ansible templating.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>