Update docker driver test for ansible 2.8 (#2050)
I also slightly modified the docker driver itself, both changes were
made with the following rationale:

Ansible 2.8 should be the default. The older versions of ansible
should be considered exceptional cases to future-proof the import,
since the new location of the HAS_DOCKER_PY constant will presumably
persist for the forseeable future.

I also removed handling for the case where HAS_DOCKER_PY cannot be
imported, under the assumption that being unable to import this constant
from either its 2.8 or pre-2.8 location is insane for all supported
versions of ansible. My apologies if this assumption is in error.

Signed-off-by: Sean Myers <sean.myers@redhat.com>