[WIP] Refactored travis.yml
* Remove sharding on functional tests

Sharding tests was hacky approach for dealing with Travis build time
limitations, one that caused lots of issues.

In order to simplify the testing and ease the migration toward other
CI system, we are better-off dropping the sharding feature (which also
seems to prevent bumping pytest).

* Remove most YAML anchors which prevented use of official syntax for
travis file for "env" blocks (lists not dictionaries).

* Consolidated newlines between jobs (readability)

Note: This is more of a POC change as we will need other optimizations
in order to assure that functional always runs under 50min time limit.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>