Backports docker_image module for internal use
As we needed extra_hosts for docker_image and that will become
available only starting with Ansible 2.9, we copy the module locally
with custom name and use it from here.

Includes note about when it will be safe to remove the internal
copy of the module.

This should not affect use of the module from tested playbooks because
its name is prefixed with `molecule_`.

Fixes: #2182

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>