WIP: Backport docker_image module for internal use (v2)
Allows users to manually enable use of newer docker_image module from
Ansible 2.9 for molecule internal docker create sequence.

This was needed in order to be able to use unreleased Ansible features,
without the risk of changing the version of Ansible used during
execution of the testing playbooks.

User can add `vendored: true` to molecule.yml to enable use of
internal backports.

That backport disables itself if current version of Ansible is already
2.9 or newer.

Background: As we needed extra_hosts for docker_image and that will
become available only starting with Ansible 2.9, we copy the module
locally with custom name and use it from here.

This should not affect Ansible version used to run tested playbooks
because its name is prefixed with `molecule_`.

There is no need to alter the tests because the vendored mode is
fully experimental and not supported. The testing with ansible-devel
should be enough to validate the execution path.

Fixes: #2182

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>