Assures ansibledevel jobs run all the time
Decision to skip running ansibledevel bought us some speed boost at
the cost of no longer testing code and introducing costly bugs.

This caused failure to release 2.22rc4 which failed on ansibledevel
jobs, which were not run during normal PRs.

From now on we need to be sure that we test all PRs with the same set of
jobs as we do for releases.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>