Makes ansible default verifier
Replace testinfra with ansible as the default verifier.

Now testinfra is an optional dependency which is installable as an
extra. This requires us to bump major version because users may need
to add this extra to keep their code compatible.

This does *not* mean we deprecate testinfra, we only transition to a
modular architecture.

Fixes: #2013
Relates: #2228
Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>