Pass merged configuration to provisioner playbooks
Up until now, molecule_yml variable in create/destroy playbooks
contained only values from the scenario molecule YAML file. This made
it a bit awkward to share common platforms definitions in different
scenarios.

Changes in this commit replace the direct access to the scenario
molecule configuration with the merged one. When the inventory and
Ansible configuration files are written to the disk, merged
configuration is also stored in the ephemeral directory of the
scenario currently being acted upon.

Signed-off-by: Tadej Borovšak <tadej.borovsak@xlab.si>