Make default Dockerfile embedded
From now on, creating a Dockerfile for docker/podman containers would
be optional. If the file is not present molecule would use its internal
version.

This ease maintenance of roles using molecule by having one less file
to create and update.

Those relying on custom containers are not affected, as this feature is
still present.

Newly created roles will not have Dockerfiles.

Signed-off-by: Sorin Sbarnea <ssbarnea@redhat.com>