Fix packaging
Removes hacks from setup.py and adds extra tests that should
prevent future regressions.

The previous workaround was broken as tox was unable to run twice
on the same environment. The assumption that newer setuptools
was preventing installation only during install was wrong, the
reality was that newer version broke packaging completely.

Fixes: #2379

Signed-off-by: Bogdan Ristea <Bogdan Ristea bogdan.ristea@gmail.com>