Avoid adding UUID twice to platform in parallel mode
Since commit 3b8c8d03, we are placing merged molecule config into the
ephemeral directory of the scenario currently being acted upon. This
configuration file already contains a preprocessed platform names if
we run molecule in parallel mode.

Because we forgot to remove the platform name mangling from the filter
that is responsible for loading the config file and exposing its
content to the playbook tasks, platform names suddenly started having
two identical suffixes.

Because those suffixes are relatively long, having two copies hit the
upper limit of the hostname length (63 characters). This caused docker
to die while trying to create new instances. The exact error message
was:

    Error starting container 123...901: 500 Server Error: Internal
    Server Error (\"OCI runtime create failed: container_linux.go:345:
    starting container process caused \"process_linux.go:430: container
    init caused \\\"sethostname: invalid argument\\\"\": unknown\")

If we remove the preprocessing of platform names from the filter
function, things go back to normal.