Remove anyconfig dependency
We used a single piece of functionality from the anyconfig library:
the dictionary merge function. In order to reduce the maintenance
burden, we decided to remove the anyconfig library and write
merge_dicts function from scratch.

We also took the opportunity and synchronized the merge behavior to
match the behavior of Ansible's combine filter, since the behavior of
the anyconfig.merge function is counter-intuitive.