Remove anyconfig dependency (#2437)
* Remove anyconfig dependency

We used a single piece of functionality from the anyconfig library:
the dictionary merge function. In order to reduce the maintenance
burden, we decided to remove the anyconfig library and write
merge_dicts function from scratch.

We also took the opportunity and synchronized the merge behavior to
match the behavior of Ansible's combine filter, since the behavior of
the anyconfig.merge function is counter-intuitive.

* Avoid mutating the input in merge_dicts function

Up until now, utils.merge_dicts function mutated its first argument
and then returned it to the caller.

In order to make this function a bit safer to use, we modified it to
operate on a shallow copy of the first argument.