tests: use is_subset to validate loaded data
Fixes recurring problem with tests which failed for users that use
DOCKER_HOST to point to a remote docker. This happened because in those
cases the loaded configuration dictionary looks different as it has
extra parameters added to it.

We failed to catch these on CI because we do not use a remote docker
connection.

From now on we should only check specific content in loaded config.

Also includes fixing of absolute imports with regression prevention
measure.