Add the ability to configure docker_networks
Currently, on creation, the docker_network module only receives the
name of the network to create. It is therefore impossible to configure
advanced details like IPAM.

Without this patch, a molecule.yaml containing multiple platforms with
multiple networks with static IPs will basically fail, as it is not possible
to determine the network ranges, because we don't pass IPAM details at the
network creation time.

It is a problem, as everyone having docker_networks should basically have
their own create/destroy playbooks to overcome this.

This fixes the behaviour by adding a new key into platforms,
named docker_networks. A new key was necessary because "networks" is used
inside the docker_container creation, which takes different arguments
than docker_network.