Improve schema validation errors
Solves bug where there where no hint regarding which data failed to validate the schema.

As molecule validates schemas of all scenarios on load, user may be confused by
a schema validation failure if the data is not displayed.

We do print the detailed data because logging is designed to work only with
single line messages. Error comes after the detail in order to improve the user experience.