Remove scenario name from molecule.yml
Scenario name is now always the folder name containing the molecule.yml and
can no longer be configured in-file.

Remove scenario.name from v2 configuration files to avoid getting an error.

This will allow us to refactor molecule, making it faster and not
forcing it to parse all scenarios in order to later run only one.

Fixes: #2434