Properly parallelize platform names
When Molecule is executed with the --parallel switch, it transforms
the platform names in order to avoid name collisions on the service
that is hosting the tests.

Currently, Molecule only parallelizes platform names and does not
touch the inventory section at all. This makes it impossible to
specify per-platform variables in the molecule.yml file.

For example, let us assume that we have this molecule.yml file:

    platforms:
      - name: my_platform
        # other stuff here

    provisioner:
      inventory:
        host_vars:
          my_platform:
            my_var: my_value

If we run such scenario, the playbook tasks have access to the my_var
variable. But as soon as we add the --parallel flag, this variable is
no longer accessible because the host names from the platforms section
and the provisioner sections do not match.

Changes in this commit ensure that when we parallelize platform names,
we also parallelize any host_vars entries that match the original
name. All other host_vars entries (the ones that do not match any of
the platform names) are left alone and are not transformed.