Mark some functional tests as extensive
Adds 'extensive' marker to some tests and configure tox to skip them
by default. The extensive tests are supposed to be indirectly
covered by the full test commands and this should decrease the CI
testing considerably.

We do not remove these tests in order to make it easy to revive
them if needed.