Make definition of docker_network future proof
Change the get_docker_networks to return a list of docker_network with
all configured options.

Move our internal defaults from the yaml file to this function and merge
it with the ones from molecule file. Internal defaults are limited to
labels and state.

Anything option defined in molecule.yml will endup being passed to the
docker_network module.

Update the tests.

Signed-off-by: Bruno Travouillon <devel@travouillon.fr>