Add mol command alias
Makes it easier to run molecule by also installing a `mol` script.
`mol` is the international abbreviation for molecule and it seems not
to be used by any other known tool.

Using the abbreviated name can be proved useful not only for terminal
usage but also when naming jobs in CI/CD pipelines, as their names
could easily become very long.