src/molecule/driver/delegated.py: Allow to set login_cmd_template in managed case
The current code only allows to configure the login command
if molecule is configured to use non managed instance but it may
be helpful to be able to set it even if the instance is managed.

For instance, if the instance (eg a chroot or a container) can be
created with the create.yml playbook, which means that this instance is
reachable with something else than ssh.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>