src/molecule/console.py: Don't always set output wrapping
The output soft wrapping with python enrich is enabled for all
terminals. This means that lines are wrapped to 80 when molecule
is run from tox or when a module ends importing this console module.
In the later case, this will wrap its output to 80 columns and
possibly break it, since the output won't be valid JSON as found in
https://github.com/ansible-community/molecule-vagrant/issues/84.

As a possible solution, enable the soft wrapping on the same condition
as for the force_terminal terminal.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>