Move logging to more standard layout
The main purpose of changes in current commit is to centralize the
logger configuration. Previously, molecule configured each and every
logger, creating new handler for each one, which made little sense.

In the new implementation, we only configure the root molecule logger.
All children logger instances inherit the configuration, making away
with duplication.

With our logging refactor, it also became feasible to use Python's
built-in logging facilities as instructed in the official docs.

Do note that we left the get_logger function in the logger module
because that piece of code is used by various molecule plugins (for
example, drivers) to obtain the properly configured logger instance.
But we again made sure we just nest the logger below the molecule root
logger and let it inherit the configuration.