Fix ANSIBLE_LIBRARY variable handling
When we introduced the ability to pass down the user-specified
ANSIBLE_LIBRARY variable value to the molecule, we wrongly assumed
that the variable will always contain a single componet, which is not
always true.

This commit makes sure that we properly handle scenarios where users
specify multiple directories in the ANSIBLE_LIBRARY environment
variable.