Change ANSIBLE_LIBRARY PATH processing
When the environment variable ANSIBLE_LIBRARY is used, molecule is
currently extending the path with the contents of the found variable,
however, this will cause problems as the system driven path will
conflict with a user defined path. To correct this issue the path
processing will now insert the user provided variable into the path.

- Bugfix Pull Request

Lanuch-Pad: https://bugs.launchpad.net/tripleo/+bug/1930757
Signed-off-by: Kevin Carter <kecarter@redhat.com>