Use cwd from BakedCommand in utils.run_command()
Fixes #151.

When trying to access Ansible vars in testinfra tests, the following
function call would fail to get values coming from host/group vars:

    ansible_vars = host.ansible('debug', 'msg={{ hostvars[inventory_hostname] }}')

This was due to: 1. testinfra being executed into the directory where
molecule was launched instead of the scenario directory ; 2. group_vars
being located in the scenario directory.

This fix set the cwd passed to subprocess_tee to the cwd set in
BakedCommand and fixes the issue.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>