Drop the use of incompatible `check-manifest` pkg
This is required because it tries to analyze `MANIFEST.in` without
understanding the context of the current packaging setup. It does not
even look into the actual sdist either.

Because of this, it makes incorrect assumptions, makes false claims
(like that some files are not included when in fact they are — via
the `setuptool-scm` integration hook). It also produces harmful
recommendations that mess up what's actually included and excluded
if they are being followed without understanding the context.