Remove selinux from being exposed as a dependency
We remove selinux from being exposed as a dependency of molecule
becauase molecule itself does not need it. It would be up to the user
to ensure that its ansible setup is correct. We still mention few
hints about selinux and virtualenvs on our installation docs.

This change will enabled us to avoid some blocking problems with
newer pip-compile where the output cannot be kept the same on linux
and non-linux.