Adapt the readthedocs config to be similar to what ansible-lint is using
This commit changes the way the python dependencies for readthedocs
are configured and mirrors the way ansible-lint is doing it. The `docs`
extra_requires is dropped and replaced with an `/docs/requirements.in`
file. The tox.ini configuration is adapted to this change.

sphinx.fail_on_warning is set to true for readthedocs config. This turns
warnings into errors (`-W` and `--keep-going options`).
Tis means the build fails if there is a warning and exits with exit
status 1. This mirrors the configuration in tox.ini.
So if building the docs locally fails, it now also fails with readthedocs.

A simple Readme.md for building documentation is added.

Also the pre-commit `pip-compile` steps are adapted, a `pip-compile-docs`
and `pip-compile-docs-upgrade` hook is added.

Fixes: #3405

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>