Use backwards incompatible change from ansible-compat 4.0.1 (#3904)
* Use run, not exec, for ansible-compat 4.0.2

4.0.2 dropped the use of `exec`, needs to be replaced with `run`

Fixes #3903

https://github.com/ansible/ansible-compat/pull/256

* bump ansible-compat 4.1.2

ansible-compat 4.0.1 included a breaking change. If we want to use
latest, minimum version must be at least 4.0.1.

ansible-compat < 4.1.2 has an issue that causes an unhandled exception
for certain log levels. https://github.com/ansible/ansible-lint/issues/3522
Since this issue caused CI trouble, we go ahead and bump to 4.1.2.

* fixup

* [DATALAD RUNCMD] recompile requirements.lock.txt

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "pip-compile --no-annotate --output-file=.config/requirements-lock.txt --resolver=backtracking --strip-extras --unsafe-package=ruamel-yaml-clib pyproject.toml",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^

* fix tests for command init role and init scenario

Currently the tests fail for `test_command_init_role` and
`test_command_init_scneario`. The tests execute `ansible-lint` on the
newly-created roles and the linter recently got more restrictive, so we
now have to adapt.

* [DATALAD RUNCMD] recompile requirements.txt

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "pip-compile --extra=docs --extra=test --no-annotate --output-file=.config/requirements.txt --resolver=backtracking --strip-extras pyproject.toml",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^

* Fake 3.9 for autogenerated files

I used 3.10, but CI will run this on 3.9, and fail if there is a diff.
This means it is safe to change this by hand, since it will be checked.

---------

Co-authored-by: Daniel Ziegenberg <daniel@ziegenberg.at>